:recycle: refactor: adjust solution folder
