+++ DsaChallenges.sln
@@ -1,0 +1,34 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{746F38C0-55E3-45C9-BCF0-DE395EEF0327}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DsaChallenges.Tests", "src\DsaChallenges.Tests\DsaChallenges.Tests.csproj", "{F9240C54-8BDF-4D56-866B-58A3929116CA}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DsaChallenges.Core", "src\DsaChallenges.Core\DsaChallenges.Core.csproj", "{D056CEE2-4698-4A30-A5DA-4EE9D56C0B2E}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{F9240C54-8BDF-4D56-866B-58A3929116CA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F9240C54-8BDF-4D56-866B-58A3929116CA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F9240C54-8BDF-4D56-866B-58A3929116CA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F9240C54-8BDF-4D56-866B-58A3929116CA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D056CEE2-4698-4A30-A5DA-4EE9D56C0B2E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D056CEE2-4698-4A30-A5DA-4EE9D56C0B2E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D056CEE2-4698-4A30-A5DA-4EE9D56C0B2E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D056CEE2-4698-4A30-A5DA-4EE9D56C0B2E}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{F9240C54-8BDF-4D56-866B-58A3929116CA} = {746F38C0-55E3-45C9-BCF0-DE395EEF0327}
+		{D056CEE2-4698-4A30-A5DA-4EE9D56C0B2E} = {746F38C0-55E3-45C9-BCF0-DE395EEF0327}
+	EndGlobalSection
+EndGlobal
