--- DsaChallenges/Program.cs
@@ -1,7 +1,0 @@
-﻿class Program
-{
-    public static void Main(string[] args)
-    {
-
-    }
-}
