+++ src/DsaChallenges.Core/LeetCode/LeetCode217.cs
@@ -1,0 +1,16 @@
+﻿namespace DsaChallenges.Core.LeetCode;
+
+public class LeetCode217
+{
+    public bool ContainsDuplicate(int[] nums)
+    {
+        var hs = new HashSet<int>();
+        foreach(int num in nums)
+        {
+            if (hs.Contains(num)) 
+                return true;
+            hs.Add(num);
+        }
+        return false;
+    }
+}
